requirements.txt
UnityPy>=1.22.5,<2.0
Pillow>=10.4.0,<11.0
numpy>=1.26.4,<3.0
trimesh>=4.5.1,<5.0
